+++ jpostman-junit/src/test/java/io/jpostman/junit/JUnitContextTest.java
@@ -132,7 +132,7 @@ public class JUnitContextTest {
 		cxt.response(executor);
 		assertEquals(cxt.statusCode(), 201);
 		assertEquals(cxt.exists("id"), true);
-		assertEquals(cxt.path("id"), 1);
+		assertEquals((Integer) cxt.path("id"), 1);
 
 		assertEquals(cxt.paths("id").toString(), "[1]");
 
@@ -226,6 +226,7 @@ public class JUnitContextTest {
 		assertEquals(third, "new-value");
 		assertEquals(calls.get(), 2);
 
+		assertEquals(secure.cache("token"), "abc123");
 		assertEquals(secure.cache().get("token"), "abc123");
 		assertEquals(secure.cache().get("cacheCanUseExplicitKeyAndReuseValue"), "new-value");